requirements.txt
Flask==3.0.0
Flask-Session==0.6.0
Werkzeug==3.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0

